+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The `pycldf` package adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## unreleased
+
+- Support certain GitHub URLs as dataset locators.
+
+
 ## [1.39.0] - 2024-09-09
 
 - Added option to `downloadmedia` subcommand to customize file naming.